Introduces concept of "type" and "list"
Configurations can be broken down in two main strategies:

* Generate one for each set X of parameters
* Generate one all combinations of a set

For and example of the former, we generate both standard platform configurations with all services, and we generate configurations for staging and demo, which include only shard.

For an example of the latter, in the case of standard configurations, we generate some common files (postinstall, logging.xml) for each service, environment (special, beta, dev) and "instance" (main service instance, b service instance).