Source path and basic main
Introduce the notion of source path for configurations, which can be used to select which files (or directories wherein they are found) will be used with a particular configuration set.

Also introduce a main which ignores the above, among other things, so that this can be tested in practice before the design is finished.